FESTOU-115: Ajustes - reautenticação + lógica de desvincular o google sendo o unico provedor

+++ lib/src/features/bottomNavBar/profile/pages/login e segurança/login_seguranca.dart
@@ -132,7 +132,7 @@ class _LoginSegurancaState extends ConsumerState<LoginSeguranca>
 
 //providerData para obter os detalhes do usuario autenticado.
   void showUserProvider() {
-    final user = FirebaseAuth.instance.currentUser!;
+    final user = FirebaseAuth.instance.currentUser;
 
     if (user != null) {
       for (var profile in user.providerData) {
@@ -145,6 +145,90 @@ class _LoginSegurancaState extends ConsumerState<LoginSeguranca>
       }
     } else {
       log("Usuário não autenticado");
+    }
+  }
+
+  Future<bool> areYouSureOnlyGoogle() async {
+    final user = FirebaseAuth.instance.currentUser;
+
+    if (user != null) {
+      bool confirmed = false;
+
+      await showDialog(
+        context: context,
+        builder: (BuildContext context) {
+          return AlertDialog(
+            title: const Text('Tem certeza?'),
+            content: const Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Text(
+                    'O google é seu único provedor, se você desvincular, sua conta será permanentemente excluída.\n'),
+                Text('Confirme que a conta é realmente a sua:\n'),
+              ],
+            ),
+            actions: <Widget>[
+              TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop(); // Fecha o diálogo
+                },
+                child: const Text('Cancelar'),
+              ),
+              TextButton(
+                onPressed: () async {
+                  Navigator.of(context).pop(); // Fecha o diálogo
+
+                  try {
+                    /*// Reautentica o usuário antes de excluir a conta
+                    AuthCredential credential = EmailAuthProvider.credential(
+                      email: user.email!,
+                      password: passwordController.text,
+                    );
+                    await user.reauthenticateWithCredential(credential);
+
+                    // Exclua a conta do usuário após a reautenticação
+                    await user.delete();*/
+
+                    //todo:
+                    /* logica para deletar o usuario que tem apenas o goole como provedor,
+                    pois: 1) se desvincular, ficará apenas o usuario lá no auth com provedor vazio, 
+                    eu queria excluir tudo ali; 2) pra deletar de fato, precisa de reautenticacao sem senha,
+                    pois só há google, nao há email/senha, entao precisaria de reautenticar com link.
+                    seria com link mesmo? há um problema em deixar o provedor la vazio?*/
+                    //?
+                    /*final authCredential = EmailAuthProvider.credentialWithLink(
+                        email: user.email!, emailLink: user.email!.toString());
+
+                    await user.linkWithCredential(authCredential);*/
+                    //?
+                    // Redirecione o usuário para a tela de login ou execute outras ações necessárias
+                    await ref
+                        .read(userFirestoreRepositoryProvider)
+                        .deleteUserDocument(user);
+                    ref.read(logoutProvider.future);
+                    log('Conta excluída com sucesso.');
+                    confirmed =
+                        true; // Indica que o usuário confirmou a exclusão
+                  } on FirebaseAuthException catch (e) {
+                    log('Erro ao excluir a conta: ${e.code}, ${e.message}');
+                    // Trate os erros conforme necessário, exiba mensagens ao usuário, etc.
+                  } catch (e) {
+                    log('Erro desconhecido ao excluir a conta: $e');
+                    // Trate outros erros conforme necessário.
+                  }
+                },
+                child: const Text('Deletar conta'),
+              ),
+            ],
+          );
+        },
+      );
+
+      return confirmed;
+    } else {
+      log('Usuario não autenticado');
+      return false;
     }
   }
 
@@ -184,48 +268,44 @@ class _LoginSegurancaState extends ConsumerState<LoginSeguranca>
               ),
               if (isUpdatingPassword) ...[
                 // Container com os campos de atualização da senha
-                Container(
-                  child: Column(
-                    children: [
-                      const TextField(
-                        decoration: InputDecoration(labelText: 'Senha Atual'),
-                        obscureText: true,
-                      ),
-                      const SizedBox(height: 0),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: [
-                          InkWell(
-                              onTap: () => Navigator.push(
-                                    context,
-                                    MaterialPageRoute(
-                                      builder: (context) =>
-                                          const EsqueciSenha(),
-                                    ),
+                Column(
+                  children: [
+                    const TextField(
+                      decoration: InputDecoration(labelText: 'Senha Atual'),
+                      obscureText: true,
+                    ),
+                    const SizedBox(height: 0),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.start,
+                      children: [
+                        InkWell(
+                            onTap: () => Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                    builder: (context) => const EsqueciSenha(),
                                   ),
-                              child: const Text('Esqueci minha senha')),
-                        ],
-                      ),
-                      const SizedBox(height: 0),
-                      const TextField(
-                        decoration: InputDecoration(labelText: 'Nova Senha'),
-                        obscureText: true,
-                      ),
-                      const SizedBox(height: 0),
-                      const TextField(
-                        decoration:
-                            InputDecoration(labelText: 'Confirmar Senha'),
-                        obscureText: true,
-                      ),
-                      const SizedBox(height: 0),
-                      ElevatedButton(
-                        onPressed: () {
-                          // Implemente a lógica para "Alterar Senha"
-                        },
-                        child: const Text('Alterar Senha'),
-                      ),
-                    ],
-                  ),
+                                ),
+                            child: const Text('Esqueci minha senha')),
+                      ],
+                    ),
+                    const SizedBox(height: 0),
+                    const TextField(
+                      decoration: InputDecoration(labelText: 'Nova Senha'),
+                      obscureText: true,
+                    ),
+                    const SizedBox(height: 0),
+                    const TextField(
+                      decoration: InputDecoration(labelText: 'Confirmar Senha'),
+                      obscureText: true,
+                    ),
+                    const SizedBox(height: 0),
+                    ElevatedButton(
+                      onPressed: () {
+                        // Implemente a lógica para "Alterar Senha"
+                      },
+                      child: const Text('Alterar Senha'),
+                    ),
+                  ],
                 ),
               ],
               const SizedBox(height: 30),
@@ -246,17 +326,27 @@ class _LoginSegurancaState extends ConsumerState<LoginSeguranca>
                                 // Verifica se o usuário está vinculado ao provedor do Google
                                 if (user.providerData.any((info) =>
                                     info.providerId == "google.com")) {
-                                  // Desvincula a conta do Google
-                                  await user.unlink("google.com");
-                                  // Atualiza a lista de provedores após o unlink
-
-                                  // Exibe uma mensagem de sucesso (pode ser ajustada conforme necessário)
-                                  ScaffoldMessenger.of(context).showSnackBar(
-                                    const SnackBar(
-                                      content: Text(
-                                          'Conta do Google desvinculada com sucesso!'),
-                                    ),
-                                  );
+                                  if (user.providerData.any((info) =>
+                                      info.providerId != "password")) {
+                                    //se desvincular, deve deletar o usuario, pois só há google como provedor.
+                                    bool confirmed =
+                                        await areYouSureOnlyGoogle();
+                                    if (confirmed) {
+                                      // Exibe uma mensagem de sucesso (pode ser ajustada conforme necessário)
+                                      ScaffoldMessenger.of(context)
+                                          .showSnackBar(
+                                        const SnackBar(
+                                          content: Text(
+                                              'Conta do Google desvinculada com sucesso!'),
+                                        ),
+                                      );
+                                    } else {
+                                      log('Usuario cancelou a açao de desvincular e deletar conta');
+                                    }
+                                  } else {
+                                    // Desvincula a conta do Google
+                                    await user.unlink("google.com");
+                                  }
                                 } else {
                                   // Caso o usuário não esteja vinculado ao provedor do Google
                                   ScaffoldMessenger.of(context).showSnackBar(
@@ -268,8 +358,7 @@ class _LoginSegurancaState extends ConsumerState<LoginSeguranca>
                                 }
                               } on FirebaseAuthException catch (e) {
                                 // Trata exceções específicas, se necessário
-                                print(
-                                    'Erro ao desvincular conta do Google: ${e.message}');
+                                log('Erro ao desvincular conta do Google, usuario nao autenticado: ${e.message}');
                               }
                             }
                           },
@@ -299,13 +388,30 @@ class _LoginSegurancaState extends ConsumerState<LoginSeguranca>
 
                     if (user != null) {
                       // Mostra um diálogo de confirmação antes de excluir a conta
-                      showDialog(
+                      await showDialog(
                         context: context,
                         builder: (BuildContext context) {
+                          TextEditingController passwordController =
+                              TextEditingController();
+
                           return AlertDialog(
-                            title: const Text('Confirmação'),
-                            content: const Text(
-                                'Tem certeza de que deseja excluir sua conta?'),
+                            title: const Text('Tem certeza?'),
+                            content: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              mainAxisSize: MainAxisSize.min,
+                              children: [
+                                const Text(
+                                    'Tem certeza de que deseja excluir sua conta?'),
+                                const Text(
+                                    'Confirme sua senha para prosseguir'),
+                                TextFormField(
+                                  controller: passwordController,
+                                  obscureText: true,
+                                  decoration:
+                                      const InputDecoration(labelText: 'Senha'),
+                                ),
+                              ],
+                            ),
                             actions: <Widget>[
                               TextButton(
                                 onPressed: () {
@@ -320,11 +426,19 @@ class _LoginSegurancaState extends ConsumerState<LoginSeguranca>
                                       .pop(); // Fecha o diálogo
 
                                   try {
-                                    // Exclua a conta do usuário
+                                    // Reautentica o usuário antes de excluir a conta
+                                    AuthCredential credential =
+                                        EmailAuthProvider.credential(
+                                      email: user.email!,
+                                      password: passwordController.text,
+                                    );
+                                    await user.reauthenticateWithCredential(
+                                        credential);
+
+                                    // Exclua a conta do usuário após a reautenticação
                                     await user.delete();
 
                                     // Redirecione o usuário para a tela de login ou execute outras ações necessárias
-
                                     await ref
                                         .read(userFirestoreRepositoryProvider)
                                         .deleteUserDocument(user);
@@ -338,7 +452,7 @@ class _LoginSegurancaState extends ConsumerState<LoginSeguranca>
                                     // Trate outros erros conforme necessário.
                                   }
                                 },
-                                child: const Text('Confirmar'),
+                                child: const Text('Deletar conta'),
                               ),
                             ],
                           );
